requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Validation and serialization
pydantic>=2.6.0,<2.10.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# HTTP client for external services
httpx==0.25.2

# Email functionality
fastapi-mail==1.4.1
jinja2==3.1.2

# System monitoring
psutil>=6.1.0

# Testing
pytest-asyncio==0.21.1


# Development tools
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
ruff==0.1.6
bandit==1.7.1
coverage==7.3.2
pytest-xdist==3.3.1
pytest-benchmark==4.0.0
pytest-sugar==0.9.7
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-timeout==2.2.0
pytest-repeat==0.9.1
pytest-randomly==3.15.0
pytest-parallel==0.1.1 